requirements.txt
Flask
PyPDF2
python-docx
fpdf
scikit-learn
werkzeug
requests
unidecode